Add desktop and AppStream metadata for Linux distros
These files can be installed by distro packagers (or users themselves
if they want system-wide integration) to add a menu entry and relevant
info in the package managers.

So far the .pro file was not edited to take these into account, so they
would have to be installed manually by packagers (together with putting
the main binary in `/usr/bin` instead of `/opt`).